Update system_prompt.txt
We are getting started with translating to Japanese. In our first test we noticed that there were some instances where no space was included between alphabets and Japanese characters including Kanji, Hiragana, and Katagana. Adding a reminder to the system prompt seems to have fixed this for us, but we have only done limited testing.